requirements.txt
#streamlit
#openai

streamlit
google-generativeai

